修改分布式锁和布隆过滤器等一系列的后续null值问题----并调试检验

+++ service/service-product/src/main/java/com/gm/gmall/product/mapper/SkuInfoMapper.java
@@ -20,7 +20,7 @@ public interface SkuInfoMapper extends BaseMapper<SkuInfo> {
 
     List<SkuSaleValueTo> getSkuSaleAndValue(Long spuId);
 
-    List<Long> getIds();
+    List<Integer> getIds();
 }
